requirements.txt
article==0.1.1
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.2
django-autoslug==1.9.6
django-ckeditor==5.8.0
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-filter==2.2.0
django-js-asset==1.2.2
django-online-users==0.3
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-taggit==0.22.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
jsonfield==3.1.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mailer==0.8.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.0
psycopg2==2.8.4
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pytz==2019.3
PyYAML==5.3
requests==2.22.0
ruamel.yaml==0.16.5
Serializer==0.2.1
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.7
wrapt==1.11.2
